feat: models 구조 변경, Search models 추가
