--- src/models/archiving/ArchivingList.ts
@@ -1,8 +1,0 @@
-export interface ArchivingListResponse {
-  [category: string]: ArchivingItem[] | []
-}
-
-export interface ArchivingItem {
-  title: string
-  contentCnt: number
-}
