--- src/models/block/Block.ts
@@ -1,3 +1,0 @@
-export interface BlockResponse {
-  nickname: string
-}
